LT-2712: Error when changing case status to Surrendered
The form library can't handle error responses that comeback from the API in the
form of a list of strings. It always expects them to be a dictionary.

This normalises API error responses into a format which will not cause an
exception and allows the error message to be shown to the user.